# CUsersvictoAppDataLocalTempopenapi.json
{"openapi":"3.1.0","info":{"title":"Watermark Remover API","version":"1.0.0"},"paths":{"/api/upload":{"post":{"summary":"Upload Video","operationId":"upload_video_api_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_video_api_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/video/{file_id}/info":{"get":{"summary":"Video Info","operationId":"video_info_api_video__file_id__info_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VideoInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/file/{file_id}/info":{"get":{"summary":"File Info","description":"Get info for any supported file type.","operationId":"file_info_api_file__file_id__info_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/video/{file_id}/stream":{"get":{"summary":"Stream Video","description":"Serve the uploaded video file with range request support for seeking.","operationId":"stream_video_api_video__file_id__stream_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/file/{file_id}/stream":{"get":{"summary":"Stream File","description":"Serve any uploaded file (image, PDF, etc.).","operationId":"stream_file_api_file__file_id__stream_get","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/video/{file_id}/detect":{"post":{"summary":"Detect Region","operationId":"detect_region_api_video__file_id__detect_post","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/file/{file_id}/detect":{"post":{"summary":"Detect File Region","description":"Auto-detect watermark region for any file type.","operationId":"detect_file_region_api_file__file_id__detect_post","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/video/{file_id}/preview":{"post":{"summary":"Preview Removal","operationId":"preview_removal_api_video__file_id__preview_post","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreviewRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/file/{file_id}/preview":{"post":{"summary":"Preview File Removal","description":"Preview watermark removal for any file type.","operationId":"preview_file_removal_api_file__file_id__preview_post","parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"string","title":"File Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreviewRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs":{"get":{"summary":"List Jobs","operationId":"list_jobs_api_jobs_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobListResponse"}}}}}},"post":{"summary":"Create Job","operationId":"create_job_api_jobs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreate"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}/cancel":{"post":{"summary":"Cancel Job","operationId":"cancel_job_api_jobs__job_id__cancel_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}/download":{"get":{"summary":"Download Job","operationId":"download_job_api_jobs__job_id__download_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/jobs/{job_id}":{"delete":{"summary":"Delete Job","operationId":"delete_job_api_jobs__job_id__delete","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_video_api_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_video_api_upload_post"},"DetectResponse":{"properties":{"region":{"$ref":"#/components/schemas/Region"},"confidence":{"type":"string","title":"Confidence","default":"auto"}},"type":"object","required":["region"],"title":"DetectResponse"},"FileInfo":{"properties":{"file_id":{"type":"string","title":"File Id"},"filename":{"type":"string","title":"Filename"},"file_type":{"$ref":"#/components/schemas/FileType"},"width":{"type":"integer","title":"Width"},"height":{"type":"integer","title":"Height"},"fps":{"type":"number","title":"Fps","default":0.0},"duration":{"type":"number","title":"Duration","default":0.0},"total_frames":{"type":"integer","title":"Total Frames","default":0},"has_audio":{"type":"boolean","title":"Has Audio","default":false},"page_count":{"type":"integer","title":"Page Count","default":0},"first_frame_base64":{"type":"string","title":"First Frame Base64","default":""}},"type":"object","required":["file_id","filename","file_type","width","height"],"title":"FileInfo"},"FileType":{"type":"string","enum":["video","image","pdf","pptx"],"title":"FileType"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobCreate":{"properties":{"file_id":{"type":"string","title":"File Id"},"region":{"$ref":"#/components/schemas/Region"},"method":{"$ref":"#/components/schemas/RemovalMethod","default":"pixel"},"sensitivity":{"type":"integer","maximum":80.0,"minimum":5.0,"title":"Sensitivity","default":30},"inpaint_radius":{"type":"integer","maximum":15.0,"minimum":1.0,"title":"Inpaint Radius","default":3},"trim_end_seconds":{"type":"number","maximum":30.0,"minimum":0.0,"title":"Trim End Seconds","default":0.0},"file_type":{"$ref":"#/components/schemas/FileType","default":"video"}},"type":"object","required":["file_id","region"],"title":"JobCreate"},"JobInfo":{"properties":{"job_id":{"type":"string","title":"Job Id"},"file_id":{"type":"string","title":"File Id"},"filename":{"type":"string","title":"Filename"},"status":{"$ref":"#/components/schemas/JobStatus"},"progress":{"type":"number","title":"Progress","default":0.0},"current_frame":{"type":"integer","title":"Current Frame","default":0},"total_frames":{"type":"integer","title":"Total Frames","default":0},"eta_seconds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Eta Seconds"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"method":{"type":"string","title":"Method","default":"pixel"}},"type":"object","required":["job_id","file_id","filename","status","created_at"],"title":"JobInfo"},"JobListResponse":{"properties":{"jobs":{"items":{"$ref":"#/components/schemas/JobInfo"},"type":"array","title":"Jobs"}},"type":"object","required":["jobs"],"title":"JobListResponse"},"JobStatus":{"type":"string","enum":["queued","processing","completed","failed","cancelled"],"title":"JobStatus"},"PreviewRequest":{"properties":{"region":{"$ref":"#/components/schemas/Region"},"method":{"$ref":"#/components/schemas/RemovalMethod","default":"pixel"},"sensitivity":{"type":"integer","maximum":80.0,"minimum":5.0,"title":"Sensitivity","default":30},"inpaint_radius":{"type":"integer","maximum":15.0,"minimum":1.0,"title":"Inpaint Radius","default":3}},"type":"object","required":["region"],"title":"PreviewRequest"},"PreviewResponse":{"properties":{"preview_base64":{"type":"string","title":"Preview Base64"},"method":{"type":"string","title":"Method"}},"type":"object","required":["preview_base64","method"],"title":"PreviewResponse"},"Region":{"properties":{"x":{"type":"integer","title":"X"},"y":{"type":"integer","title":"Y"},"w":{"type":"integer","title":"W"},"h":{"type":"integer","title":"H"}},"type":"object","required":["x","y","w","h"],"title":"Region"},"RemovalMethod":{"type":"string","enum":["pixel","box","crop","delete_shapes"],"title":"RemovalMethod"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VideoInfo":{"properties":{"file_id":{"type":"string","title":"File Id"},"filename":{"type":"string","title":"Filename"},"width":{"type":"integer","title":"Width"},"height":{"type":"integer","title":"Height"},"fps":{"type":"number","title":"Fps"},"duration":{"type":"number","title":"Duration"},"total_frames":{"type":"integer","title":"Total Frames"},"has_audio":{"type":"boolean","title":"Has Audio"},"first_frame_base64":{"type":"string","title":"First Frame Base64","default":""}},"type":"object","required":["file_id","filename","width","height","fps","duration","total_frames","has_audio"],"title":"VideoInfo"}}}}